Fixed categories page UX issues
Issues reported after testing:
1. Edit Category button on detail view did nothing and had poor contrast.
2. Assigned products took full rows — wanted chips so many fit per row.
3. Product count showed 0 in list view even after assigning products.

Fixes:

1. Edit/Delete dialogs were rendered ONLY inside the list-view branch of the
   component. When the detail view returned early, the dialogs weren't in
   the DOM — opening them from detail had no effect. Refactored: the
   dialog JSX is now defined once as a shared 'dialogs' variable and
   rendered at the end of BOTH view returns. Edit button now uses the solid
   brand-blue CTA style for clearer affordance.

2. Assigned products list layout changed from divide-y full-width rows to
   flex-wrap pill/chip style — each chip shows product name + HS code +
   remove button. Many products now fit per row, scaling naturally.

3. Product/container counts replaced SQL correlated subqueries with two
   separate GROUP BY count queries, mapped into lookup tables and merged
   into each category row. Postgres COUNT(*) is now ::int cast so the value
   comes through as a number. More importantly, refreshCategories() is now
   called when the user clicks the back arrow (returning to list view),
   ensuring the most recent assignment data is shown.

+++ app/api/admin/products/route.ts
@@ -26,7 +26,7 @@ export async function GET() {
                 lastSyncedAt: products.lastSyncedAt,
                 updatedAt: products.updatedAt,
                 containerCount: sql<number>`
-                    (SELECT COUNT(*) FROM ${containers}
+                    (SELECT COUNT(*)::int FROM ${containers}
                      WHERE ${containers.categoryId} = ${products.categoryId}
                      AND ${containers.status} IN ('OPEN','THRESHOLD_REACHED'))
                 `.as("container_count"),
